Move methods from FileUtils to Formula
These don't need to live on FileUtils and don't really make sense there
either.
