+++ Library/Homebrew/formula.rb
@@ -1907,6 +1907,56 @@ class Formula
     end
   end
 
+  # A version of `FileUtils.mkdir` that also changes to that folder in
+  # a block.
+  def mkdir(name)
+    result = FileUtils.mkdir_p(name)
+    return result unless block_given?
+    FileUtils.chdir name do
+      yield
+    end
+  end
+
+  # Run `scons` using a Homebrew-installed version rather than whatever is
+  # in the `PATH`.
+  def scons(*args)
+    system Formulary.factory("scons").opt_bin/"scons", *args
+  end
+
+  # Run `make` 3.81 or newer.
+  # Uses the system make on Leopard and newer, and the
+  # path to the actually-installed make on Tiger or older.
+  def make(*args)
+    if Utils.popen_read("/usr/bin/make", "--version")
+            .match(/Make (\d\.\d+)/)[1] > "3.80"
+      make_path = "/usr/bin/make"
+    else
+      make = Formula["make"].opt_bin/"make"
+      make_path = if make.exist?
+        make.to_s
+      else
+        (Formula["make"].opt_bin/"gmake").to_s
+      end
+    end
+
+    if superenv?
+      make_name = File.basename(make_path)
+      with_env(HOMEBREW_MAKE: make_name) do
+        system "make", *args
+      end
+    else
+      system make_path, *args
+    end
+  end
+
+  # Run `xcodebuild` without Homebrew's compiler environment variables set.
+  def xcodebuild(*args)
+    removed = ENV.remove_cc_etc
+    system "xcodebuild", *args
+  ensure
+    ENV.update(removed)
+  end
+
   private
 
   # Returns the prefix for a given formula version number.
